Fixed a type of maxcc parameter in opensearch utility function

+++ sentinelhub/api/opensearch.py
@@ -92,7 +92,7 @@ def get_area_info(bbox: BBox, date_interval: RawTimeIntervalType, maxcc: Optiona
     return list(result_list)
 
 
-def get_area_dates(bbox: BBox, date_interval: RawTimeIntervalType, maxcc: Optional[int] = None) -> List[dt.date]:
+def get_area_dates(bbox: BBox, date_interval: RawTimeIntervalType, maxcc: Optional[float] = None) -> List[dt.date]:
     """Get list of times of existing images from specified area and time range
 
     :param bbox: bounding box of requested area
